lesson_x/client/requirements.txt
chardet
requests
pytest
pytest-cov
fabric3
pyyaml
tabulate
sqlalchemy
pycryptodome
pycryptodomex
PyQt5
